+++ vck/src/commonTest/kotlin/at/asitplus/wallet/lib/agent/AgentSdJwtTest.kt
@@ -19,7 +19,7 @@ import com.benasher44.uuid.uuid4
 import io.github.aakira.napier.Napier
 import io.kotest.core.spec.style.FreeSpec
 import io.kotest.matchers.collections.shouldBeSingleton
-import io.kotest.matchers.collections.shouldHaveSingleElement
+import io.kotest.matchers.collections.shouldContain
 import io.kotest.matchers.collections.shouldHaveSize
 import io.kotest.matchers.shouldBe
 import io.kotest.matchers.types.shouldBeInstanceOf
@@ -67,12 +67,10 @@ class AgentSdJwtTest : FreeSpec({
 
         val verified = verifier.verifyPresentation(vp.sdJwt, challenge)
             .shouldBeInstanceOf<Verifier.VerifyPresentationResult.SuccessSdJwt>()
-        verified.reconstructed.claims shouldHaveSize 2
+        verified.reconstructedJsonObject.entries shouldHaveSize 2
 
-        verified.reconstructed.claims.first { it.claimName == CLAIM_GIVEN_NAME }
-            .claimValue.jsonPrimitive.content shouldBe "Susanne"
-        verified.reconstructed.claims.first { it.claimName == CLAIM_DATE_OF_BIRTH }
-            .claimValue.jsonPrimitive.content shouldBe "1990-01-01"
+        verified.reconstructedJsonObject[CLAIM_GIVEN_NAME]?.jsonPrimitive?.content shouldBe "Susanne"
+        verified.reconstructedJsonObject[CLAIM_DATE_OF_BIRTH]?.jsonPrimitive?.content shouldBe "1990-01-01"
         verified.isRevoked shouldBe false
     }
 
@@ -88,8 +86,8 @@ class AgentSdJwtTest : FreeSpec({
         ).sdJwt
         val verified = verifier.verifyPresentation(sdJwt, challenge)
             .shouldBeInstanceOf<Verifier.VerifyPresentationResult.SuccessSdJwt>()
-        verified.reconstructed.claims.shouldBeSingleton()
-        verified.reconstructed.claims.shouldHaveSingleElement { it.claimName == CLAIM_GIVEN_NAME }
+        verified.reconstructedJsonObject.entries.shouldBeSingleton()
+        verified.reconstructedJsonObject.keys shouldContain CLAIM_GIVEN_NAME
         verified.isRevoked shouldBe false
     }